FDL widget added to the llrf expert GUI
The FDL widget to control the Fast Data Logger for the Loops and the
Diagnostics cards has been added to the LLRF GUI.
There is missing the FDL Gui to analyze the data. For that reason the
buttons to open this gui has been disable.

Also some test must be done before this branch can be attached to the
master branch.